move state from input component to Product
